move adreess label reducers to data/bitcoin
