+++ packages/blockchain-wallet-v4-frontend/src/data/rootReducer.js
@@ -41,8 +41,7 @@ const rootReducer = combineReducers({
   [paths.walletPath]: coreReducers.wallet,
   [paths.settingsPath]: coreReducers.settings,
   [paths.walletOptionsPath]: coreReducers.walletOptions,
-  [paths.kvStorePath]: coreReducers.kvStore,
-  [paths.addressLabelsPath]: coreReducers.addressLabels
+  [paths.kvStorePath]: coreReducers.kvStore
 })
 
 export default rootReducer
