+++ packages/blockchain-wallet-v4/src/redux/actions.js
@@ -1,4 +1,3 @@
-import * as addressLabels from './addressLabels/actions.js'
 import * as data from './data/actions.js'
 import * as wallet from './wallet/actions.js'
 import * as settings from './settings/actions.js'
@@ -9,7 +8,6 @@ import * as kvStore from './kvStore/actions.js'
 import * as refresh from './refresh/actions.js'
 
 export {
-  addressLabels,
   data,
   kvStore,
   settings,
