--- packages/blockchain-wallet-v4/src/redux/addressLabels/actionTypes.js
@@ -1,1 +1,0 @@
-export const ADD_ADDRESS_LABEL = '@CORE.ADD_ADDRESS_LABEL'
