--- packages/blockchain-wallet-v4/src/redux/addressLabels/actions.js
@@ -1,3 +1,0 @@
-import * as AT from './actionTypes'
-
-export const addAddressLabel = (address, label) => ({type: AT.ADD_ADDRESS_LABEL, payload: {address, label}})
