--- packages/blockchain-wallet-v4/src/redux/addressLabels/selectors.js
@@ -1,4 +1,0 @@
-import { path } from 'ramda'
-import {addressLabelsPath} from '../paths'
-
-export const getAddressLabel = (address) => path([addressLabelsPath, address])
