+++ packages/blockchain-wallet-v4/src/redux/data/bitcoin/actionTypes.js
@@ -36,3 +36,5 @@ export const FETCH_BITCOIN_TRANSACTION_HISTORY_LOADING = '@CORE.FETCH_BITCOIN_TR
 export const FETCH_BITCOIN_TRANSACTION_HISTORY_SUCCESS = '@CORE.FETCH_BITCOIN_TRANSACTION_HISTORY_SUCCESS'
 export const FETCH_BITCOIN_TRANSACTION_HISTORY_FAILURE = '@CORE.FETCH_BITCOIN_TRANSACTION_HISTORY_FAILURE'
 export const CLEAR_BITCOIN_TRANSACTION_HISTORY = '@CORE.CLEAR_BITCOIN_TRANSACTION_HISTORY'
+
+export const ADD_ADDRESS_LABEL = '@CORE.ADD_ADDRESS_LABEL'
