+++ packages/blockchain-wallet-v4/src/redux/paths.js
@@ -3,4 +3,3 @@ export const walletPath = 'walletPath'
 export const kvStorePath = 'kvStorePath'
 export const settingsPath = 'settingsPath'
 export const walletOptionsPath = 'walletOptionsPath'
-export const addressLabelsPath = 'addressLabelsPath'
