+++ packages/blockchain-wallet-v4/src/redux/reducers.js
@@ -3,13 +3,11 @@ import wallet from './wallet/reducers.js'
 import settings from './settings/reducers.js'
 import walletOptions from './walletOptions/reducers.js'
 import kvStore from './kvStore/reducers.js'
-import addressLabels from './addressLabels/reducers.js'
 
 export {
   data,
   wallet,
   settings,
   walletOptions,
-  kvStore,
-  addressLabels
+  kvStore
 }
